PWA-2231 Moved the hook function to redux helpers.
